--- Assets/Scenes/MainMenu.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2442d9d546f51164684bd689ea40b239
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
