--- Assets/Scenes/StudentCard.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: da713ffc3dde0794dabf04b2fc5d9316
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
