--- Assets/Texts/Dialog.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<dialogue>
-	<node>
-		<text>Великий феодал, в предыдущем этапе вам и вашим соратникам удалось восстановить «Священное знание». Но это была лишь пустая оболочка. Теперь же необходимо пробудить его, пройдя испытание.</text>
-	</node>
-	<node>
-		<text>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et
-		dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-		Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non 
-		proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</text>
-	</node>
-	<node end="true">
-		<text>А теперь заполним студак</text>
-	</node>
-
-	<node>
-		<text>Для начала примерим на себя роль адепта.</text>
-	</node>
-	<node>
-		<text>Будучи адептом, вам необходимо собрать одну из двух комбинаций осколков, представленных в нижней части экрана.</text>
-	</node>
-
-
-	<node end="true">
-		<text>Осколки складываются в выделенные для них ячейки.</text>
-	</node>
-
-
-	<node>
-		<text>Сами осколки для комбинаций подбираются по ходу игры из алтаря. Коснитесь алтаря, чтобы взять свой первый осколок.</text>
-	</node>
-
-
-	<node>
-		<text>Максимально можно хранить столько же осколков, сколько есть ячеек под них. Если вы возьмете лишний, то он заменит один случайный из уже имеющихся.</text>
-	</node>
-
-
-	<node>
-		<text>Когда вы берете осколок, у вас повышается уровень эфирного стресса, а когда его берет кто-то другой — снижается.</text>
-	</node>
-
-
-	<node>
-		<text>Если вы достигните максимального уровня стресса, то не сможете снова брать осколки, пока значение не опустится до нуля.</text>
-	</node>
-	<node end="true">
-		<text>Чтобы быстрее снизить эфирный стресс, а также получить другую помощь, адепты могут попросить магистров использовать заклинание поддержки.</text>
-	</node>
-</dialogue>
