+++ Assets/Texts/Dialogue.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<dialogue>
+    <node>
+        <text>Добро пожаловать! Рад видеть, что ты заинтересовался нашей магистратурой.</text>
+    </node>
+    <node end="true">
+        <text>Для начала давай заведём тебе студенческий билет — впиши своё имя, как хочешь, чтобы мы тебя называли.
+        </text>
+    </node>
+    <node>
+        <text>Отлично, теперь немного разомнём мозги.</text>
+    </node>
+    <node>
+        <text>Найди в таблице ключевые слова, которые связаны с нашей программой. Уверен, справишься быстро!</text>
+    </node>
+    <node>
+        <text>А теперь важный момент — выбери направление МагоЛего, которое тебе ближе всего.</text>
+    </node>
+    <node>
+        <text>От этого выбора будет зависеть, какой диплом ты получишь в конце.</text>
+    </node>
+    <node>
+        <text>Молодец! Главное не просто знать правильный ответ, а понимать, почему он такой.</text>
+    </node>
+    <node>
+        <text>У нас в магистратуре ты научишься именно этому.</text>
+    </node>
+    <node>
+        <text>Поздравляю с окончанием игры!</text>
+    </node>
+    <node>
+        <text>
+            Твой путь только начинается, и я надеюсь, что мы встретимся снова — уже в аудиториях магистратуры ЦИКИ.
+        </text>
+    </node>
+</dialogue>
