+++ Assets/UI/DialogSystem/DialogPanel.png.meta
@@ -160,7 +160,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      DialogPanel_0: -8032391750203985770
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
