+++ Assets/UI/DialogSystem/dialogMan.png.meta
@@ -160,7 +160,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      dialogMan_0: -5426764648773263683
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
